First Commit
The Telegram Bot Email Sender is a Python-based project that enables users to send emails through a Telegram bot. The bot acts as an interface, allowing users to interactively provide recipient email addresses, email subject, email body, and even upload images as attachments. The project utilizes the telebot library for Telegram bot functionality and the smtplib library for sending emails.